Server-role: Complete CRUD Role and Permission

+++ blog-app-server/src/main/java/com/example/blog_app/services/implement/UserServiceImpl.java
@@ -1,10 +1,14 @@
 package com.example.blog_app.services.implement;
 
 import com.example.blog_app.exceptions.DuplicateResourceException;
+import com.example.blog_app.exceptions.ImmutableResourceException;
 import com.example.blog_app.exceptions.ResourceNotFoundException;
+import com.example.blog_app.models.dtos.RoleDto;
 import com.example.blog_app.models.dtos.UserRequestDto;
 import com.example.blog_app.models.dtos.UserResponseDto;
+import com.example.blog_app.models.entities.Role;
 import com.example.blog_app.models.entities.User;
+import com.example.blog_app.repositories.RoleRepository;
 import com.example.blog_app.repositories.UserRepository;
 import com.example.blog_app.services.UserService;
 import org.modelmapper.ModelMapper;
@@ -18,11 +22,13 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
     private final BCryptPasswordEncoder passwordEncoder;
     private final ModelMapper modelMapper;
 
-    public UserServiceImpl(UserRepository userRepository, BCryptPasswordEncoder passwordEncoder, ModelMapper modelMapper) {
+    public UserServiceImpl(UserRepository userRepository, RoleRepository roleRepository, BCryptPasswordEncoder passwordEncoder, ModelMapper modelMapper) {
         this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
         this.passwordEncoder = passwordEncoder;
         this.modelMapper = modelMapper;
     }
@@ -40,9 +46,13 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserResponseDto updateUser(UserRequestDto userDto, Integer userId) {
+    public UserResponseDto updateUser(UserRequestDto userDto, Long userId) {
         User user = this.userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
+
+        if(user.isImmutable()) {
+            throw new ImmutableResourceException("User immutable");
+        }
 
         user.setName(userDto.getName());
         user.setEmail(userDto.getEmail());
@@ -56,7 +66,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserResponseDto getUser(Integer userId) {
+    public UserResponseDto getUser(Long userId) {
         User user = this.userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
         return this.userToDto(user);
@@ -71,10 +81,44 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Integer userId) {
+    public void deleteUser(Long userId) {
         User user = this.userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
+
+        if(user.isImmutable()) {
+            throw new ImmutableResourceException("User immutable");
+        }
+
         this.userRepository.delete(user);
+    }
+
+    @Override
+    public List<RoleDto> getUserRoles(Long userId) {
+        User user = this.userRepository.findById(userId).
+                orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
+        return user.getRoles().stream()
+                .map(role -> modelMapper.map(role, RoleDto.class))
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public void assignRoleToUser(Long userId, String roleName){
+        User user = this.userRepository.findById(userId).
+                orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
+        Role role = this.roleRepository.findByName(roleName).
+                orElseThrow(() -> new ResourceNotFoundException("Role not found with ID: " + roleName));
+        user.getRoles().add(role);
+        this.userRepository.save(user);
+    }
+
+    @Override
+    public void unassignRoleFromUser(Long userId, String roleName) {
+        User user = this.userRepository.findById(userId).
+                orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
+        Role role = this.roleRepository.findByName(roleName).
+                orElseThrow(() -> new ResourceNotFoundException("Role not found with ID: " + roleName));
+        user.getRoles().remove(role);
+        this.userRepository.save(user);
     }
 
     private User dtoToUser(UserRequestDto userDto) {
